+++ aufs.go
@@ -32,25 +32,12 @@ import (
 
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/mount"
-	"github.com/containerd/containerd/platforms"
-	"github.com/containerd/containerd/plugin"
 	"github.com/containerd/containerd/snapshots"
 	"github.com/containerd/containerd/snapshots/storage"
 	"github.com/containerd/continuity/fs"
 	"github.com/pkg/errors"
 )
 
-func init() {
-	plugin.Register(&plugin.Registration{
-		Type: plugin.SnapshotPlugin,
-		ID:   "aufs",
-		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			ic.Meta.Platforms = append(ic.Meta.Platforms, platforms.DefaultSpec())
-			ic.Meta.Exports["root"] = ic.Root
-			return New(ic.Root)
-		},
-	})
-}
 
 var (
 	dirperm        sync.Once
@@ -65,7 +52,7 @@ type snapshotter struct {
 // New creates a new snapshotter using aufs
 func New(root string) (snapshots.Snapshotter, error) {
 	if err := supported(); err != nil {
-		return nil, errors.Wrap(plugin.ErrSkipPlugin, err.Error())
+		return nil, err
 	}
 	if err := os.MkdirAll(root, 0700); err != nil {
 		return nil, err
